refactor: standardize folder structure and namespaces
