+++ InfiniteSquaresWebAPI/Interfaces/IMappingService.cs
@@ -1,10 +1,11 @@
 ﻿using InfiniteSquaresCore.Models;
 using InfiniteSquaresWebAPI.DTOs;
 
-namespace InfiniteSquaresWebAPI.Interface;
+namespace InfiniteSquaresWebAPI.Interfaces;
 
 public interface IMappingService
 {
     SquareDto MapToDto(Square square);
     Square MapToEntity(SquareDto squareDto);
 }
+
